Recharge page now separate generation class
